requirement.txt
fastapi==0.110.0
uvicorn==0.29.0
transformers==4.38.0
torch==2.2.0
pillow==10.2.0